requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
reportlab
langchain
langchain-core
langchain-groq
openai
tqdm
